+++ packages/config/src/chains.ts
@@ -452,7 +452,7 @@ export const darwiniaCrab = new EvmParachain({
   parachainId: 2105,
   rpc: 'https://crab-rpc.darwinia.network',
   ss58Format: 18,
-  ws: ['wss://crab-rpc.darwinia.network', 'wss://darwiniacrab-rpc.dwellir.com'],
+  ws: ['wss://darwiniacrab-rpc.dwellir.com', 'wss://crab-rpc.darwinia.network'],
 });
 
 export const fantomTestnet = new EvmChain({
@@ -1083,6 +1083,9 @@ export const moonbeam = new EvmParachain({
     ChainAsset.fromAsset(fil, {
       address: '0xfFFfFFFF6C57e17D210DF507c82807149fFd70B2',
       decimals: 18,
+      ids: {
+        id: '144012926827374458669278577633504620722',
+      },
     }),
     ChainAsset.fromAsset(glmr, {
       address: '0x0000000000000000000000000000000000000802',