+++ packages/config/src/xcm-configs/alphanetAssetHub.ts
@@ -29,7 +29,9 @@ export const alphanetAssetHubRoutes = new ChainRoutes({
         chain: moonbaseAlpha,
         balance: BalanceBuilder().substrate().assets().account(),
         fee: {
-          amount: FeeBuilder().assetManager().assetTypeUnitsPerSecond(),
+          amount: FeeBuilder()
+            .xcmPaymentApi()
+            .xcmPaymentFee({ isAssetReserveChain: false }),
           asset: tt1,
         },
       },