+++ packages/config/src/xcm-configs/alphanetRelay.ts
@@ -2,10 +2,10 @@ import {
   BalanceBuilder,
   ExtrinsicBuilder,
   FeeBuilder,
-} from '@moonbeam-network/xcm-builder';
-import { unit } from '../assets';
-import { alphanetRelay, moonbaseAlpha } from '../chains';
-import { ChainRoutes } from '../types/ChainRoutes';
+} from "@moonbeam-network/xcm-builder";
+import { unit } from "../assets";
+import { alphanetRelay, moonbaseAlpha } from "../chains";
+import { ChainRoutes } from "../types/ChainRoutes";
 
 export const alphanetRelayRoutes = new ChainRoutes({
   chain: alphanetRelay,
@@ -23,7 +23,9 @@ export const alphanetRelayRoutes = new ChainRoutes({
         chain: moonbaseAlpha,
         balance: BalanceBuilder().substrate().assets().account(),
         fee: {
-          amount: FeeBuilder().assetManager().assetTypeUnitsPerSecond(),
+          amount: FeeBuilder()
+            .xcmPaymentApi()
+            .xcmPaymentFee({ isAssetReserveChain: false }),
           asset: unit,
         },
       },