+++ packages/config/src/xcm-configs/khala.ts
@@ -23,7 +23,9 @@ export const khalaRouts = new ChainRoutes({
         chain: moonriver,
         balance: BalanceBuilder().substrate().assets().account(),
         fee: {
-          amount: FeeBuilder().assetManager().assetTypeUnitsPerSecond(),
+          amount: FeeBuilder()
+            .xcmPaymentApi()
+            .xcmPaymentFee({ isAssetReserveChain: false }),
           asset: pha,
         },
       },
@@ -46,7 +48,9 @@ export const khalaRouts = new ChainRoutes({
         chain: moonriver,
         balance: BalanceBuilder().substrate().system().account(),
         fee: {
-          amount: 0.001,
+          amount: FeeBuilder()
+            .xcmPaymentApi()
+            .xcmPaymentFee({ isAssetReserveChain: true }),
           asset: movr,
         },
       },